% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAll.R
\name{all.plot}
\alias{all.plot}
\title{all.plot}
\usage{
\method{all}{plot}(path = NA)
}
\arguments{
\item{path}{sets the path to cnv files. Defaults to 'NA' (and "data/")}
}
\description{
Function to plot all chromosomes for all samples in defiuned path
}
\examples{
all.plot(path="data/")
}
\keyword{all}
\keyword{plot}
