% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAll.R
\name{allPlot}
\alias{allPlot}
\title{allPlot}
\usage{
allPlot(path = NA)
}
\arguments{
\item{path}{Sets the path to cnv files. [Default "data/"]}
}
\description{
Function to plot all chromosomes for all samples in path provided
}
\examples{
allPlot(path="data/")
}
\keyword{all}
\keyword{plot}
