% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChrom.R
\name{chrom.plot}
\alias{chrom.plot}
\title{chrom.plot}
\usage{
chrom.plot(chrom = NA, cnv_file, ylim = c(-5, 5))
}
\arguments{
\item{chrom}{Specify the chromosome to plot (defualts to 'X')}

\item{cnv_file}{cnv file to plot. Required}

\item{ylim}{allows adjustment of y axis limits in plot}
}
\description{
Function to plot a single chromosomes for a given sample
}
\examples{
chrom.plot(cnv_file = "data/HUM-7.tagged.SC.hits.filt-vs-HUM-9.tagged.SC.hits.filt.window-10000.minw-4.cnv")
}
\keyword{chrom}
\keyword{plot}
