% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{cleanTheme}
\alias{cleanTheme}
\title{cleanTheme}
\usage{
cleanTheme(base_size = 12)
}
\arguments{
\item{base_size}{= sets the base font size, defaults to 12}
}
\description{
Clean theme for plotting
}
\examples{
cleanTheme()
}
\keyword{theme}
