% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileClean.R
\name{file.cleanR}
\alias{file.cleanR}
\title{file.cleanR}
\usage{
file.cleanR(cnv_file, region = F)
}
\arguments{
\item{cnv_file}{cnv file to clean. Required}

\item{region}{more stringent filter for larger windows as defualt}
}
\description{
Function to clean the cnv files
}
\examples{
file.cleanR()
}
\keyword{clean}
