% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freecPlot.R
\name{freecPlot}
\alias{freecPlot}
\title{FreecPlot}
\usage{
freecPlot(cnv_file = "data/freec/HUM-7.tagged.filt.SC.RG.bam_ratio.txt")
}
\arguments{
\item{cnv_file}{File to plot. [Required]}
}
\description{
Function to plot the corrected CN for Control Freec "ratio" files.
}
\examples{
freecPlot(cnv_file = "data/freec/HUM-7.tagged.filt.SC.RG.bam_ratio.txt")
}
\keyword{CN}
\keyword{Control}
\keyword{Freec}
\keyword{file}
\keyword{for}
\keyword{normalised}
\keyword{plot}
\keyword{ratio}
