% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAll.R
\name{plot.all.grid}
\alias{plot.all.grid}
\title{plot.all.grid}
\usage{
\method{plot}{all.grid}(path = NA)
}
\arguments{
\item{path}{sets the path to cnv files. Defaults to 'NA' (and "data/")}
}
\description{
Function to plot all chromosomes for all samples in defiuned path
}
\examples{
plot.all.grid(path="data/")
plot.all.grid()
}
\keyword{all}
\keyword{plot}
