% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoverage.R
\name{plotCoverage}
\alias{plotCoverage}
\title{plotCoverage}
\usage{

  plotCoverage(counts_file = "data/counts/A785-A788R11.tagged.SC.hits.filt-vs-A785-A788R12.tagged.SC.hits.filt.window-50000.minw-4.count",
  readLength = 100, windowSize = 10000)
}
\arguments{
\item{counts_file}{File to plot. [Required]}

\item{readLength}{Specifiy the length of read [Default 100]}

\item{windowSize}{Specifiy the window size used [Default 50000]}
}
\description{
Function to plot raw coverage for a tumour/normal sample
}
\examples{
plotCoverage(counts_file = 'data/counts/A785-A788R11.tagged.SC.hits.filt-vs-A785-A788R12.tagged.SC.hits.filt.window-50000.minw-4.count', readLength=100,windowSize = 50000)
}
\keyword{chrom}
\keyword{plot}
