% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegion.R
\name{region.plot}
\alias{region.plot}
\title{region.plot}
\usage{
region.plot(cnv_file, from = NA, to = NA, chrom = NA, ylim = c(-5, 5),
  tick = 1e+06, bp1 = NA, bp2 = NA, title = NA)
}
\arguments{
\item{cnv_file}{cnv file to plot. Required}

\item{from}{start coordinate of plotting window}

\item{to}{end coordinate of plotting window}

\item{chrom}{Specify the chromosome to plot (defualts to 'X')}

\item{ylim}{allows adjustment of y axis limits in plot}

\item{tick}{allows adjustment of spacing of ticks. Defualt 1Mb}

\item{bp1}{draw a vertical line at given coordinate to mark bp1}

\item{bp2}{draw a vertical line at given coordinate to mark bp2}

\item{title}{provide a title for the plot}
}
\description{
Function to plot a particular region from a single chromosomes for a given sample
}
\examples{
region.plot(cnv_file="data/w500/HUM-7.tagged.SC.hits.filt-vs-HUM-9.tagged.SC.hits.filt.window-500.minw-4.cnv", from=3050000, to=3450000, chrom="X", ylim=c(-7,7), bp1=3129368,bp2=3352041, tick=100000, title="222Kb DEL on X")
}
\keyword{plot}
\keyword{region}
