% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegion.R
\name{regionPlot}
\alias{regionPlot}
\title{regionPlot}
\usage{
regionPlot(cnv_file, from = NA, to = NA, chrom = NA, ylim = c(-5, 5),
  tick = 100000, bp1 = NA, bp2 = NA, title = NA)
}
\arguments{
\item{cnv_file}{File to plot. [Required]}

\item{from}{Start coordinate of plotting window [Default from=2950000 ]}

\item{to}{End coordinate of plotting window [Default to=3400000]}

\item{chrom}{Specify the chromosome to plot [Default 'X']}

\item{ylim}{Adjust y axis limits in plot [Default ylim=c(-5,5)]}

\item{tick}{Adjust spacing of ticks [Defualt 1Mb]}

\item{bp1}{Draw a vertical line at given coordinate to mark bp1}

\item{bp2}{Draw a vertical line at given coordinate to mark bp2}

\item{title}{Plot title}
}
\description{
Function to plot a particular region from a single chromosomes for a sample
}
\examples{
regionPlot(cnv_file="data/w500/test.window-500.cnv", from=3050000, to=3450000, chrom="X", ylim=c(-7,7), bp1=3129368,bp2=3352041, tick=100000, title="222Kb DEL on X")
}
\keyword{plot}
\keyword{region}
