% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAll.R
\name{allPlot}
\alias{allPlot}
\title{allPlot}
\usage{
allPlot(path = NA)
}
\arguments{
\item{path}{sets the path to cnv files. Defaults to 'NA' (and "data/")}
}
\description{
Function to plot all chromosomes for all samples in defiuned path
}
\examples{
allPlot(path="data/")
}
\keyword{all}
\keyword{plot}
