% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChrom.R
\name{chromPlot}
\alias{chromPlot}
\title{chromPlot}
\usage{
chromPlot(chrom = NA, cnv_file, ylim = c(-5, 5))
}
\arguments{
\item{chrom}{Specify the chromosome to plot (defualts to 'X')}

\item{cnv_file}{cnv file to plot. Required}

\item{ylim}{allows adjustment of y axis limits in plot}
}
\description{
Function to plot a single chromosomes for a given sample
}
\examples{
chromPlot(cnv_file = "data/test.window-10000.cnv")
}
\keyword{chrom}
\keyword{plot}
