% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChrom.R
\name{plot.chrom}
\alias{plot.chrom}
\title{plot.chrom}
\usage{
plot.chrom(chrom = NA, cnv_file, ylim = c(-5, 5))
}
\arguments{
\item{chrom}{Specify the chromosome to plot (defualts to 'X')}

\item{cnv_file}{cnv file to plot. Required}

\item{ylim}{allows adjustment of y axis limits in plot}
}
\description{
Function to plot a single chromosomes for a given sample
}
\examples{
plot.chrom()
}
\keyword{chrom}
\keyword{plot}
